CI/CD,Docs/pyproject: Bump Python and requirements to 3.10 - 3.14
With the EOL of Python 3.9, and release of Python 3.14, we bump the
range of supported Pythons version in the CI/CD and the documentation.
We also bump the requirements in the uv.lock and pyproject.toml file to
match this range. If dependencies did not need to be updated (e.g.
because their version already supported Python 3.14), they have been
left untouched.

Note that this commit does not yet include any other changes, e.g.
modernization of the codebase due to relying on 3.10+.

Required dependency updates in pyproject.toml, due to no wheels being
available for Python 3.10 for versions prior to this:
- numpy: >= 1.20.0 -> >= 1.21.2
- scipy: >= 1.6.0 -> >= 1.7.2

Required dependency updates to uv.lock:
- casadi v3.7.0 -> v3.7.2
- cftime v1.6.4.post1 -> v1.6.5
- numpy v2.2.6 -> v2.2.6, v2.3.5
- pyyaml v6.0.2 -> v6.0.3
- scipy v1.15.3 -> v1.15.3, v1.16.3

Note that we have some new version splits appearing (different
dependencies in the lock file for different versions of Python), due to
e.g. NumPy 2.3.5 not having binary wheels for Python 3.10.

New version splits:

  contourpy (was already split, but updated):
    1.3.2                <- py < 3.11
    1.3.3                <- py >= 3.12

  numpy:
    2.2.6                <- py < 3.11
    2.3.5                <- py >= 3.12

  scipy:
    1.15.3               <- py < 3.11
    1.16.3               <- py >= 3.12

Signed-off-by: Jack Vreeken <jack@vreeken.me>