SimulationProblem: Fix NaN warning potential symbol name mismatch
With symbols possibly being larger than 1, the length of __state_vector
and __sym_list ar not necessarily equal. For example, we do the
following for delay expressions:

    expression_symbol = ca.MX.sym(expression_state, n_previous_values)

We therefore need to use the indices in the sym_dict to properly match a
certain state vector entry to its symbol name.

Note that this issue popped up when trying to make sure all zip()
statements can accept the 'strict=True' argument.

Signed-off-by: Jack Vreeken <jack@vreeken.me>